Added post editing feature for users
